+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - content page import validation using the model
+- Add Age question type for forms
 
 ### Changed
 - Increased SMS limit to 459 characters